revert integration test
