+++ integration_tests/pyproject.toml
@@ -28,7 +28,6 @@ eth-bloom = "^3.0"
 eth-hash = "^0"
 pyyaml = "^6.0.2rc1"
 pytest-timeout = "^2.3.1"
-rpds-py = "0.17.1"
 
 [tool.poetry.dev-dependencies]
